bot recuperacion/requirements.txt
discord.py==2.6.4
aiomysql==0.3.2
python-dotenv==1.2.1
